Update main layout for more dynamic material design feel
Rewrote most of the activity_article_list.xml layout file so that the appbar on the main activity feels more dynamic and in line with material design. The logo now moves up with a scroll upwards until it hits a ceiling and and pins at the standard appbar size. It will hold in that position until the user scrolls the RecylerView to its top again and then the appbar will extend back to its maximum size. Thanks to the CollapsingToolbarLayout, the elevation is handled dynamically as well with scrolling states.